Made your IPeripheralProvider registration multiloader rather than per-platform
