+++ common/src/main/kotlin/net/spaceeye/PlatformUtils.kt
@@ -1,0 +1,11 @@
+package net.spaceeye
+
+import dan200.computercraft.api.peripheral.IPeripheralProvider
+import dev.architectury.injectables.annotations.ExpectPlatform
+
+object PlatformUtils {
+    @ExpectPlatform
+    fun getPeripheralProvider(): IPeripheralProvider {
+        throw AssertionError()
+    }
+}
