+++ fabric/src/main/java/net/spaceeye/someperipherals/fabric/PlatformUtilsImpl.java
@@ -1,0 +1,10 @@
+package net.spaceeye.someperipherals.fabric;
+
+import dan200.computercraft.api.peripheral.IPeripheralProvider;
+import net.spaceeye.someperipherals.fabric.integrations.cc.SomePeripheralsPeripheralProviderFabric;
+
+public class PlatformUtilsImpl {
+    public static IPeripheralProvider getPeripheralProvider() {
+        return new SomePeripheralsPeripheralProviderFabric();
+    }
+}
